+++ lib/matchers/validations/length_of.rb
@@ -18,11 +18,18 @@ module Mongoid
         end
         alias :with_max :with_maximum
 
+        def within(value)
+          @within = value
+          self
+        end
+        alias :in :within
+
         def matches?(subject)
           return false unless @result = super(subject)
 
           check_minimum if @minimum
           check_maximum if @maximum
+          check_range   if @within
 
           @result
         end
@@ -31,6 +38,7 @@ module Mongoid
           desc = []
           desc << " with minimum #{@minimum}" if @minimum
           desc << " with maximum #{@maximum}" if @maximum
+          desc << " within range #{@within}"  if @within
           super << desc.to_sentence
         end
 
@@ -52,6 +60,19 @@ module Mongoid
             @positive_message << " with maximum of #{actual}"
           else
             @negative_message << " with maximum of #{actual}"
+            @result = false
+          end
+        end
+
+        def check_range
+          min, max = [@within.min, @within.max]
+          actual_min = @validator.options[:minimum]
+          actual_max = @validator.options[:maximum]
+
+          if actual_min == min && actual_max == max
+            @positive_message << " with range #{@within}"
+          else
+            @negative_message << " with range #{actual_min..actual_max}"
             @result = false
           end
         end