+++ test/models/models.rb
@@ -19,6 +19,7 @@ class Person
   validates_uniqueness_of(:email)
   validates_uniqueness_of(:login, scope: :site, case_sensitive: false)
   validates_length_of(:password, minimum: 8, maximum: 16)
+  validates_length_of(:login, in: 5..12)
   validates_format_of(:email, with: /^([^\s]+)((?:[-a-z0-9]\.)[a-z]{2,})$/i)
   validates_inclusion_of(:role, in: ["admin", "user"])
   validates_exclusion_of(:email, in: ["foo@bar.com", "fizz@buzz.com"])