+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.0.2
 
++ Added `in(range)` alias to `within` matcher.
++ Added `within(range)` chain matcher to `validate_length_of`.
 + Added `scoped_to(*scope)` chain matcher to `validate_uniqueness_of`
 + Added `with_message(custom_message)` chain matcher for validation matchers to test custom messages.
 + Added `DocumentModuleMatcher` class to simplify document matchers.